+++ docs/changelog.md
@@ -6,6 +6,7 @@
 
 - Add filter options to sub-command sync #46
 - Add check sub-command to validate mani config
+- Add option to disable spinner when running tasks
 
 ## 0.21.0